requirements.txt
asgiref==3.6.0
Django==4.1.9
django-admin-thumbnails==0.2.6
django-appconf==1.0.5
django-ckeditor==6.5.1
django-compressor==4.3
django-extensions==3.2.1
django-filter==22.1
django-htmx==1.13.0
django-js-asset==2.0.0
gunicorn==20.1.0
Pillow==9.3.0
psycopg==3.1.8
psycopg2==2.9.5
python-decouple==3.6
python-dotenv==0.21.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
rcssmin==1.1.1
rjsmin==1.2.1
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.5.0
tzdata==2022.6
tzlocal==4.2
whitenoise==6.2.0
